Formatting + adding events to segment updating.
