+++ packages/client/src/pages/FlowViewerv2/SidePanel/viewers/MessageViewer.tsx
@@ -81,18 +81,18 @@ const MessageViewer: FC<SidePanelComponentProps<MessageNodeData>> = ({
                   }
                 >
                   <Tooltip
-                      placement="top-start"
-                      className="bg-[#111827] rounded-none p-[5px] text-white"
-                      content={`${nodeData.stats?.[stat.key] || 0}`}
+                    placement="top-start"
+                    className="bg-[#111827] rounded-none p-[5px] text-white"
+                    content={`${nodeData.stats?.[stat.key] || 0}`}
                   >
-                  <div className="font-roboto text-[14px] leading-[22px] text-[#6B7280]">
-                    {stat.name}
-                  </div>
+                    <div className="font-roboto text-[14px] leading-[22px] text-[#6B7280]">
+                      {stat.name}
+                    </div>
                   </Tooltip>
                   <div className="font-roboto text-[24px] leading-[28px]">
-                      <div className="relative">
-                        <div className="w-[8px] h-[8px] bg-transparent rounded-[100%]" />
-                      </div>
+                    <div className="relative">
+                      <div className="w-[8px] h-[8px] bg-transparent rounded-[100%]" />
+                    </div>
                     {stat.renderLabel(nodeData.stats?.[stat.key] || 0)}
                   </div>
                 </div>
