+++ packages/server/src/api/steps/processors/time.window.step.processor.ts
@@ -57,7 +57,7 @@ export class TimeWindowStepProcessor extends WorkerHost {
     private journeyLocationsService: JourneyLocationsService,
     @Inject(StepsService) private stepsService: StepsService,
     @Inject(CacheService) private cacheService: CacheService,
-    @Inject(QueueService) private queueService: QueueService,
+    @Inject(QueueService) private queueService: QueueService
   ) {
     super();
   }
@@ -202,7 +202,8 @@ export class TimeWindowStepProcessor extends WorkerHost {
           );
 
           if (nextStep) {
-            const nextStepDepth: number = this.queueService.getNextStepDepthFromJob(job);
+            const nextStepDepth: number =
+              this.queueService.getNextStepDepthFromJob(job);
 
             if (
               nextStep.type !== StepType.TIME_DELAY &&
@@ -217,7 +218,7 @@ export class TimeWindowStepProcessor extends WorkerHost {
                 customer: job.data.customer,
                 location: job.data.location,
                 event: job.data.event,
-                stepDepth: nextStepDepth
+                stepDepth: nextStepDepth,
               };
             } else {
               // Destination is time based,
