+++ packages/server/src/api/steps/processors/wait.until.step.processor.ts
@@ -51,7 +51,7 @@ export class WaitUntilStepProcessor extends WorkerHost {
     private journeyLocationsService: JourneyLocationsService,
     @Inject(StepsService) private stepsService: StepsService,
     @Inject(CacheService) private cacheService: CacheService,
-    @Inject(QueueService) private queueService: QueueService,
+    @Inject(QueueService) private queueService: QueueService
   ) {
     super();
   }
@@ -213,7 +213,8 @@ export class WaitUntilStepProcessor extends WorkerHost {
             );
 
             if (nextStep) {
-              const nextStepDepth: number = this.queueService.getNextStepDepthFromJob(job);
+              const nextStepDepth: number =
+                this.queueService.getNextStepDepthFromJob(job);
 
               if (
                 nextStep.type !== StepType.TIME_DELAY &&
@@ -228,7 +229,7 @@ export class WaitUntilStepProcessor extends WorkerHost {
                   customer: job.data.customer,
                   location: job.data.location,
                   event: job.data.event,
-                  stepDepth: nextStepDepth
+                  stepDepth: nextStepDepth,
                 };
               } else {
                 // Destination is time based,
