+++ packages/server/src/api/steps/steps.module.ts
@@ -46,6 +46,7 @@ import { TimeDelayStepProcessor } from './processors/time.delay.step.processor';
 import { TimeWindowStepProcessor } from './processors/time.window.step.processor';
 import { WaitUntilStepProcessor } from './processors/wait.until.step.processor';
 import { QueueService } from '@/common/services/queue.service';
+import { SegmentsModule } from '../segments/segments.module';
 
 function getProvidersList() {
   let providerList: Array<any> = [
@@ -54,7 +55,7 @@ function getProvidersList() {
     RedlockService,
     JourneyLocationsService,
     CacheService,
-    QueueService
+    QueueService,
   ];
 
   if (process.env.LAUDSPEAKER_PROCESS_TYPE == 'QUEUE') {
@@ -143,6 +144,7 @@ function getProvidersList() {
     forwardRef(() => JourneysModule),
     forwardRef(() => OrganizationsModule),
     forwardRef(() => WorkspacesModule),
+    forwardRef(() => SegmentsModule),
     SlackModule,
   ],
   providers: getProvidersList(),
