+++ packages/server/src/api/tests/tests.service.ts
@@ -369,7 +369,7 @@ export class TestsService {
   public async getSegmentSize(segmentId: string, session: string) {
     return await this.segmentCustomersRepository.count({
       where: {
-        segment: segmentId, //{id: segmentId,},
+        segment: { id: segmentId },
       },
     });
   }
