+++ packages/server/src/api/webhooks/webhooks.service.ts
@@ -22,7 +22,7 @@ import { InjectQueue } from '@nestjs/bullmq';
 import { Queue } from 'bullmq';
 import { Webhook } from 'svix';
 import fetch from 'node-fetch'; // Ensure you have node-fetch if you're using Node.js
-import { ProviderType } from '../events/events.preprocessor';
+import { ProviderType } from '../events/processors/events.preprocessor';
 import { Organization } from '../organizations/entities/organization.entity';
 import {
   DEFAULT_PLAN,
